requirements.txt
# Core dependencies
groq>=0.4.0,<1.0.0  # Major version compatibility
python-dotenv>=0.19.0  # Stable LTS version
pydantic>=2.0.0,<3.0.0  # Major version 2.x

# Async support
aiohttp>=3.8.0  # For async HTTP requests

# Frontend
streamlit>=1.32.0  # For web interface

# Document processing
python-docx>=0.8.11  # For Word documents
PyPDF2>=3.0.0  # For PDF files
python-magic-bin>=0.4.14  # For MIME type detection (Windows) 